refactor: modify mixin classes for deltaRCM to separate concerns.
All methods necessary for the operation of pyDeltaRCM were contained in 
a single Tools class, which rendered the entire thing pretty unreadable. 
To improve legibility, I separated out the methods related to water 
routing, sediment routing, and grid initialization. I also made a class 
for methods shared by water and sediment routing. 

A next step is to just reorder the methos so that they correspond more 
closely with the order that the methods are called, and add some more 
commenting. This probably falls under documentation, so I'll say that 
this closes #iss11.