feat: separate BMI wrapper from deltaRCM model
In this commit, I went through. and tried to find all the things in the 
core model that are wrapped and accessed by the BMI wrapper. In the 
process, I tried to take things that didn't need to be in the wrapper, 
and instead put them in the main model. The biggest thing like this that 
I found were the model time, which is critical for the model, but was 
defined both in the BMI wrapper and in the core model itself. The other 
thing was a list of input and output variable names, that is used to 
parse the yaml file that specifies the run. 

I moved all these things into the core model, under the premise that if 
the model is instantiated from a BMI framework, it will acquire those 
processes once it is initialized, and doesn't need them before 
initialization.