feat: provide yaml workflow for BMI wrapper.
Institute a new workflow, where a yaml can be specified with the 
BMI-compliant names, which is then parsed using the enclosed dictionary, 
and saved as a temporary model-compliant yaml file, which is fed into 
the model.

Dictionary values declaring types were switched to be python data types 
for the parsing comparison. These types were syncronized across the 
model and the bmi wrapper. tests and other scripts now point to the 
test_bmi.yaml