refactor: change way that weighted random choice made for water
As originally written, the code would calculate the weighted probability 
of stepping to a given cell on the fly, meaning that this computation 
happened every time parcels stepped forward. This change moves the 
computation of the probabilities to the front of the operation, meaning 
that before any water parcels route, we pre-compute all the random walk 
probabilities. 

I also incorporated elements of the random walk choice making itself, 
which turned out to be an expensive procedure. Now, the 
`get_weight_at_cell()` function should return a flattened 1D array that 
contains the cumulative weights. this array is reshaped to be a 3x3 
array giving the window around the current cell.