feat: add command line hook for deltaRCM
I agree with Andrew, that the command line would be a good way to 
interface with the model, especially if configuration files could be 
passed from the command line.

In this commit, I add that feature using a console_scripts entrypoint in 
setup.py to a special new function called run_model, and this function 
runs the model for a given number of timesteps. This number of timesteps 
is set by the config file, which now accepts an additional timesteps 
parameter.