tests: add unit tests for new command line function
Basically, since we don't get a model instance to work with when we test 
from command line, I just hacked this together to save an output image, 
and check if that image exists. This is enabled with a new 
test_output.yaml file.

I also added a test to make sure the versions match when we query from 
command line.