fix save_grids()
Changes save_grids() function to log and raise an error if the
grid saving to the netCDF file does not work as expected.

Also adds a test to check this functionality by initializing a
model without a netCDF file and then trying to write a grid
to the non-existant netcdf file.